+++ pysat/instruments/cnofs_ivm.py
@@ -58,7 +58,7 @@ import numpy as np
 
 import pysat
 
-from .methods import nasa_cdaweb as cdw
+from pysat.instruments.methods import nasa_cdaweb as cdw
 
 platform = 'cnofs'
 name = 'ivm'