+++ pysat/instruments/timed_see.py
@@ -41,7 +41,7 @@ from __future__ import absolute_import
 import functools
 
 import pysat
-from .methods import nasa_cdaweb as cdw
+from pysat.instruments.methods import nasa_cdaweb as cdw
 
 # include basic instrument info
 platform = 'timed'