+++ CHANGELOG.md
@@ -20,7 +20,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - Fixed boolean logic in when checking for start and stop dates in `_instrument.download`
   - Fixed loading of COSMIC atmPrf files
   - Fixed feedback from COSMIC GPS when data not found on remote server
+  - Fixed deprecation warning for pysat.utils.coords.scale_units
   - Fixed a bug when trying to combine empty f107 lists
+  - Made import of methods more robust
 
 ## [2.1.0] - 2019-11-18
 - New Features